+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 7.6-SNAPSHOT
 
 ADDED:
 
+- new KeyFactory algorithm 'RSASSA-PSS' added to the enum KeyPairGeneratorAlgorithm
 - new unit tests created for verify the values of the enums
 
 Version 7.5